Changed all instances of v1::RenderOperation::OperationType for OperationType Ported PSOs to D3D11. WIP. Compiles but doesn't work properly yet. PSOs made a lot of v1 code obsolete (D3D11HLSLProgram::getLayoutForVao, overriding D3D11VertexDeclaration, RenderSystem::setVertexDeclaration & RenderSystem::setVertexBufferBinding) OpenGL on Windows still not compiling.

+++ Components/Hlms/Unlit/src/OgreHlmsUnlit.cpp
@@ -429,7 +429,7 @@ namespace Ogre
         passCache.passPso = getPassPsoForScene( sceneManager );
         passCache.properties = mSetProperties;
 
-        assert( mPassCache.size() <= HlmsBits::PassMask &&
+        assert( mPassCache.size() <= (size_t)HlmsBits::PassMask &&
                 "Too many passes combinations, we'll overflow the bits assigned in the hash!" );
         PassCacheVec::iterator it = std::find( mPassCache.begin(), mPassCache.end(), passCache );
         if( it == mPassCache.end() )
